requirements.txt
pelican[markdown]
gitpython
typogrify
